% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_info.R
\name{nfl_get_team_info}
\alias{nfl_get_team_info}
\title{get the links for each nfl team}
\usage{
nfl_get_team_info(file)
}
\arguments{
\item{file}{The location to store a data frame with the link for each team.}
}
\description{
get the links for each nfl team
}
